+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Add `Apply` methods which runs `ApplyDirectly`, `ApplyCustomResources` and `ApplyDeployments`.
 - Add command `applier apply ...` which call the `Apply` method.
 - The `files` parameter in `Apply`, `ApplyDirectly`, `ApplyDeployments` and `ApplyCustomResources` can be an array of directory or file rather then only files.
+- Add `--header` in the command-lines.
+- Add `--excluded` to exclude files included by the `--path`
 
 ## Breaking changes